requirements.txt
aiofiles==23.2.1
aiogram==3.7.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
APScheduler==3.10.4
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.6.2
fake-useragent==1.5.1
frozenlist==1.4.1
idna==3.7
magic-filter==1.0.12
multidict==6.0.5
pydantic==2.7.3
pydantic_core==2.18.4
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
soupsieve==2.5
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
yarl==1.9.4
